+++ tests/linalg_test.py
@@ -18,6 +18,7 @@ from functools import partial
 import unittest
 
 import numpy as np
+import scipy
 import scipy as osp
 
 from absl.testing import absltest
@@ -1410,6 +1411,35 @@ class ScipyLinalgTest(jtu.JaxTestCase):
         return jsp.linalg.expm(x, upper_triangular=False, max_squarings=16)
       jtu.check_grads(expm, (a,), modes=["fwd", "rev"], order=1, atol=tol,
                       rtol=tol)
+
+class EighTridiagonalTest(jtu.JaxTestCase):
+
+  @parameterized.named_parameters(jtu.cases_from_list(
+     {"testcase_name": f"_n={n}_dtype={dtype.__name__}",
+      "n": n, "dtype": dtype}
+     for n in [1, 2, 3, 7, 8, 100]
+     for dtype in float_types + complex_types))
+  def testToeplitz(self, n, dtype):
+    jtu.skip_if_unsupported_type(dtype)
+    for a, b in [[2, -1], [1, 0], [0, 1], [-1e10, 1e10], [-1e-10, 1e-10]]:
+      alpha = a * np.ones([n], dtype=dtype)
+      beta = b * np.ones([n - 1], dtype=dtype)
+      # scipy.linalg.eigh_tridiagonal doesn't support complex inputs, so for
+      # this we call the slower numpy.linalg.eigh.
+      if np.issubdtype(alpha.dtype, np.complexfloating):
+        tridiagonal = np.diag(alpha) + np.diag(beta, 1) + np.diag(
+            np.conj(beta), -1)
+        eigvals_expected, _ = np.linalg.eigh(tridiagonal)
+      else:
+        eigvals_expected = scipy.linalg.eigh_tridiagonal(
+            alpha, beta, eigvals_only=True)
+      eigvals = jax.scipy.linalg.eigh_tridiagonal(
+          alpha, beta, eigvals_only=True)
+      finfo = np.finfo(dtype)
+      atol = 4 * np.sqrt(n) * finfo.eps * np.amax(np.abs(eigvals_expected))
+      self.assertAllClose(eigvals_expected, eigvals, atol=atol, rtol=1e-4)
+
+
 
 if __name__ == "__main__":
   absltest.main(testLoader=jtu.JaxTestLoader())